Move tree utility class under util
