+++ frontend/src/components/Editor.tsx
@@ -9,7 +9,8 @@ import 'react-markdown-editor-lite/lib/index.css';
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../app/hooks';
 import { APIStatus, APIStatusType } from '../reducer/common';
-import { getNoteAsync, resetStatus, saveNoteAsync, selectNoteStatus, selectNotesTree, TreeUtil } from '../reducer/noteSlice';
+import { getNoteAsync, resetStatus, saveNoteAsync, selectNoteStatus, selectNotesTree } from '../reducer/noteSlice';
+import TreeUtil from '../util/TreeUtil';
 import { appendPath, getDecodedPath, getFilenameFromTitle, getSanitizedErrorMessage, getTitleFromFilename, splitPath } from '../util/util';
 import CustomReactMarkdown from './lib/CustomReactMarkdown';
